+++ bitcoin/src/bip32.rs
@@ -12,14 +12,14 @@ use core::fmt;
 use core::ops::Index;
 use core::str::FromStr;
 
-use bitcoin_internals::write_err;
+use bitcoin_internals::{impl_array_newtype, write_err};
 use secp256k1::{self, Secp256k1, XOnlyPublicKey};
 #[cfg(feature = "serde")]
 use serde;
 
 use crate::hash_types::XpubIdentifier;
 use crate::hashes::{hex, sha512, Hash, HashEngine, Hmac, HmacEngine};
-use crate::internal_macros::{impl_array_newtype, impl_bytes_newtype, serde_string_impl};
+use crate::internal_macros::{impl_bytes_newtype, serde_string_impl};
 use crate::io::Write;
 use crate::network::constants::Network;
 use crate::prelude::*;
